services/entity_extraction/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
deeppavlov==1.7.0
exceptiongroup==1.3.1
fastapi==0.89.1
filelock==3.9.1
fsspec==2025.10.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.2.0
huggingface_hub==0.36.2
idna==3.11
joblib==1.5.3
nltk==3.9.2
numpy==1.23.5
packaging==26.0
pandas==1.5.3
prometheus_client==0.24.1
protobuf==3.20.0
pybind11==2.10.3
pydantic==1.10.26
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
pytorch-crf==0.7.2
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.0.2
scipy==1.10.0
sentencepiece==0.2.0
six==1.17.0
starlette==0.22.0
threadpoolctl==3.6.0
tokenizers==0.13.3
torch==1.13.1
TorchCRF==1.1.0
tqdm==4.64.1
transformers==4.30.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.18.3
